+++ demo/hello_world/run.py
@@ -1,7 +1,9 @@
 import gradio as gr
 
+
 def greet(name):
     return "Hello " + name + "!"
+
 
 demo = gr.Interface(fn=greet, inputs="textbox", outputs="textbox")