+++ gradio/components/login_button.py
@@ -5,7 +5,8 @@ from __future__ import annotations
 import json
 import time
 import warnings
-from typing import TYPE_CHECKING, Literal, Sequence
+from collections.abc import Sequence
+from typing import TYPE_CHECKING, Literal
 
 from gradio_client.documentation import document
 
@@ -32,7 +33,7 @@ class LoginButton(Button):
         *,
         every: Timer | float | None = None,
         inputs: Component | Sequence[Component] | set[Component] | None = None,
-        variant: Literal["primary", "secondary", "stop"] = "secondary",
+        variant: Literal["primary", "secondary", "stop", "huggingface"] = "huggingface",
         size: Literal["sm", "lg"] | None = None,
         icon: str
         | None = "https://huggingface.co/front/assets/huggingface_logo-noborder.svg",
@@ -43,18 +44,13 @@ class LoginButton(Button):
         elem_classes: list[str] | str | None = None,
         render: bool = True,
         key: int | str | None = None,
-        scale: int | None = 0,
+        scale: int | None = None,
         min_width: int | None = None,
-        signed_in_value: str = "Signed in as {}",
     ):
         """
         Parameters:
             logout_value: The text to display when the user is signed in. The string should contain a placeholder for the username with a call-to-action to logout, e.g. "Logout ({})".
         """
-        if signed_in_value != "Signed in as {}":
-            warnings.warn(
-                "The `signed_in_value` parameter is deprecated. Please use `logout_value` instead."
-            )
         self.logout_value = logout_value
         super().__init__(
             value,