+++ gradio/processing_utils.py
@@ -2,16 +2,20 @@ from __future__ import annotations
 
 import base64
 import hashlib
+import ipaddress
 import json
 import logging
 import os
 import shutil
+import socket
 import subprocess
 import tempfile
 import warnings
+from functools import lru_cache
 from io import BytesIO
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
+from urllib.parse import urlparse, urlunparse
 
 import aiofiles
 import httpx
@@ -20,9 +24,11 @@ from gradio_client import utils as client_utils
 from PIL import Image, ImageOps, ImageSequence, PngImagePlugin
 
 from gradio import utils, wasm_utils
+from gradio.context import LocalContext
 from gradio.data_classes import FileData, GradioModel, GradioRootModel, JsonData
-from gradio.exceptions import Error
-from gradio.utils import abspath, get_upload_folder, is_in_or_equal
+from gradio.exceptions import Error, InvalidPathError
+from gradio.route_utils import API_PREFIX
+from gradio.utils import abspath, get_hash_seed, get_upload_folder, is_in_or_equal
 
 with warnings.catch_warnings():
     warnings.simplefilter("ignore")  # Ignore pydub warning if ffmpeg is not installed
@@ -34,11 +40,11 @@ if wasm_utils.IS_WASM:
 
     # NOTE: In the Wasm env, we use urllib3 to make HTTP requests. See https://github.com/gradio-app/gradio/issues/6837.
     class Urllib3ResponseSyncByteStream(httpx.SyncByteStream):
-        def __init__(self, response) -> None:
+        def __init__(self, response: urllib3.HTTPResponse) -> None:
             self.response = response
 
         def __iter__(self):
-            yield from self.response.stream()
+            yield from self.response.stream(decode_content=True)
 
     class Urllib3Transport(httpx.BaseTransport):
         def __init__(self):
@@ -58,16 +64,22 @@ if wasm_utils.IS_WASM:
                 preload_content=False,  # Stream the content
             )
 
+            # HTTPX's gzip decoder sometimes fails to decode the content in the Wasm env as https://github.com/gradio-app/gradio/pull/9333#issuecomment-2348048882,
+            # so we avoid it by removing the content-encoding header passed to httpx.Response,
+            # and handle the decoding in `Urllib3ResponseSyncByteStream.__iter__()` with `urllib3`'s implementation.
+            response_headers = response.headers.copy()
+            response_headers.discard("content-encoding")
+
             return httpx.Response(
                 status_code=response.status,
-                headers=response.headers,
+                headers=response_headers,
                 stream=Urllib3ResponseSyncByteStream(response),
             )
 
     sync_transport = Urllib3Transport()
 
     class PyodideHttpResponseAsyncByteStream(httpx.AsyncByteStream):
-        def __init__(self, response) -> None:
+        def __init__(self, response: pyodide.http.FetchResponse) -> None:
             self.response = response
 
         async def __aiter__(self):
@@ -177,8 +189,12 @@ def encode_pil_to_bytes(pil_image, format="png"):
         return output_bytes.getvalue()
 
 
+hash_seed = get_hash_seed().encode("utf-8")
+
+
 def hash_file(file_path: str | Path, chunk_num_blocks: int = 128) -> str:
     sha = hashlib.sha256()
+    sha.update(hash_seed)
     with open(file_path, "rb") as f:
         for chunk in iter(lambda: f.read(chunk_num_blocks * sha.block_size), b""):
             sha.update(chunk)
@@ -187,18 +203,21 @@ def hash_file(file_path: str | Path, chunk_num_blocks: int = 128) -> str:
 
 def hash_url(url: str) -> str:
     sha = hashlib.sha256()
+    sha.update(hash_seed)
     sha.update(url.encode("utf-8"))
     return sha.hexdigest()
 
 
 def hash_bytes(bytes: bytes):
     sha = hashlib.sha256()
+    sha.update(hash_seed)
     sha.update(bytes)
     return sha.hexdigest()
 
 
 def hash_base64(base64_encoding: str, chunk_num_blocks: int = 128) -> str:
     sha = hashlib.sha256()
+    sha.update(hash_seed)
     for i in range(0, len(base64_encoding), chunk_num_blocks * sha.block_size):
         data = base64_encoding[i : i + chunk_num_blocks * sha.block_size]
         sha.update(data.encode("utf-8"))
@@ -260,9 +279,80 @@ def save_file_to_cache(file_path: str | Path, cache_dir: str) -> str:
     return full_temp_file_path
 
 
+@lru_cache(maxsize=256)
+def resolve_with_google_dns(hostname: str) -> str | None:
+    url = f"https://dns.google/resolve?name={hostname}&type=A"
+
+    if wasm_utils.IS_WASM:
+        import pyodide.http
+
+        content = pyodide.http.open_url(url)
+        data = json.load(content)
+    else:
+        import urllib.request
+
+        with urllib.request.urlopen(url) as response:
+            data = json.loads(response.read().decode())
+
+    if data.get("Status") == 0 and "Answer" in data:
+        for answer in data["Answer"]:
+            if answer["type"] == 1:
+                return answer["data"]
+
+
+# Always return these URLs as is, without checking to see if they resolve
+# to an internal IP address. This is because Hugging Face uses DNS splitting,
+# which means that requests from HF Spaces to HF Datasets or HF Models
+# may resolve to internal IP addresses even if they are publicly accessible.
+PUBLIC_URL_WHITELIST = ["hf.co", "huggingface.co"]
+
+
+def get_public_url(url: str) -> str:
+    parsed_url = urlparse(url)
+    if parsed_url.scheme not in ["http", "https"]:
+        raise httpx.RequestError(f"Invalid scheme for URL: {url}")
+    hostname = parsed_url.hostname
+    if not hostname:
+        raise httpx.RequestError(f"Invalid URL: {url}, missing hostname")
+    if hostname.lower() in PUBLIC_URL_WHITELIST:
+        return url
+
+    try:
+        addrinfo = socket.getaddrinfo(hostname, None)
+    except socket.gaierror as e:
+        raise httpx.RequestError(
+            f"Cannot resolve URL with hostname: {hostname}, please download this file and use the path instead."
+        ) from e
+
+    for family, _, _, _, sockaddr in addrinfo:
+        ip = sockaddr[0]
+        if family == socket.AF_INET6:
+            ip = ip.split("%")[0]  # Remove scope ID if present
+
+        if ipaddress.ip_address(ip).is_global:
+            return url
+
+    google_resolved_ip = resolve_with_google_dns(hostname)
+    if google_resolved_ip and ipaddress.ip_address(google_resolved_ip).is_global:
+        if parsed_url.scheme == "https":
+            return url
+        new_parsed = parsed_url._replace(netloc=google_resolved_ip)
+        if parsed_url.port:
+            new_parsed = new_parsed._replace(
+                netloc=f"{google_resolved_ip}:{parsed_url.port}"
+            )
+        return urlunparse(new_parsed)
+
+    raise httpx.RequestError(
+        f"No public IP address found for URL: {url}, please download this file and use the path instead."
+    )
+
+
 def save_url_to_cache(url: str, cache_dir: str) -> str:
     """Downloads a file and makes a temporary file path for a copy if does not already
     exist. Otherwise returns the path to the existing temp file."""
+    url = get_public_url(url)
+
     temp_dir = hash_url(url)
     temp_dir = Path(cache_dir) / temp_dir
     temp_dir.mkdir(exist_ok=True, parents=True)
@@ -270,10 +360,14 @@ def save_url_to_cache(url: str, cache_dir: str) -> str:
     full_temp_file_path = str(abspath(temp_dir / name))
 
     if not Path(full_temp_file_path).exists():
-        with sync_client.stream("GET", url, follow_redirects=True) as r, open(
-            full_temp_file_path, "wb"
-        ) as f:
-            for chunk in r.iter_raw():
+        with (
+            sync_client.stream("GET", url, follow_redirects=True) as response,
+            open(full_temp_file_path, "wb") as f,
+        ):
+            for redirect in response.history:
+                get_public_url(str(redirect.url))
+
+            for chunk in response.iter_raw():
                 f.write(chunk)
 
     return full_temp_file_path
@@ -282,6 +376,8 @@ def save_url_to_cache(url: str, cache_dir: str) -> str:
 async def async_save_url_to_cache(url: str, cache_dir: str) -> str:
     """Downloads a file and makes a temporary file path for a copy if does not already
     exist. Otherwise returns the path to the existing temp file. Uses async httpx."""
+    url = get_public_url(url)
+
     temp_dir = hash_url(url)
     temp_dir = Path(cache_dir) / temp_dir
     temp_dir.mkdir(exist_ok=True, parents=True)
@@ -290,6 +386,9 @@ async def async_save_url_to_cache(url: str, cache_dir: str) -> str:
 
     if not Path(full_temp_file_path).exists():
         async with async_client.stream("GET", url, follow_redirects=True) as response:
+            for redirect in response.history:
+                get_public_url(str(redirect.url))
+
             async with aiofiles.open(full_temp_file_path, "wb") as f:
                 async for chunk in response.aiter_raw():
                     await f.write(chunk)
@@ -380,14 +479,8 @@ def move_files_to_cache(
             pass
         elif not block.proxy_url:
             # If the file is on a remote server, do not move it to cache.
-            if check_in_upload_folder and not client_utils.is_http_url_like(
-                payload.path
-            ):
-                path = os.path.abspath(payload.path)
-                if not is_in_or_equal(path, get_upload_folder()):
-                    raise ValueError(
-                        f"File {path} is not in the upload folder and cannot be accessed."
-                    )
+            if not client_utils.is_http_url_like(payload.path):
+                _check_allowed(payload.path, check_in_upload_folder)
             if not payload.is_stream:
                 temp_file_path = block.move_resource_to_block_cache(payload.path)
                 if temp_file_path is None:
@@ -396,14 +489,16 @@ def move_files_to_cache(
                 if keep_in_cache:
                     block.keep_in_cache.add(payload.path)
 
-        url_prefix = "/stream/" if payload.is_stream else "/file="
+        url_prefix = (
+            f"{API_PREFIX}/stream/" if payload.is_stream else f"{API_PREFIX}/file="
+        )
         if block.proxy_url:
             proxy_url = block.proxy_url.rstrip("/")
-            url = f"/proxy={proxy_url}{url_prefix}{payload.path}"
+            url = f"{API_PREFIX}/proxy={proxy_url}{url_prefix}{payload.path}"
         elif client_utils.is_http_url_like(payload.path) or payload.path.startswith(
             f"{url_prefix}"
         ):
-            url = payload.path
+            url = f"{payload.path}"
         else:
             url = f"{url_prefix}{payload.path}"
         payload.url = url
@@ -414,6 +509,52 @@ def move_files_to_cache(
         data = data.model_dump()
 
     return client_utils.traverse(data, _move_to_cache, client_utils.is_file_obj)
+
+
+def _check_allowed(path: str | Path, check_in_upload_folder: bool):
+    blocks = LocalContext.blocks.get()
+    if blocks is None or not blocks.has_launched:
+        return
+
+    abs_path = utils.abspath(path)
+
+    created_paths = [utils.get_upload_folder()]
+    # if check_in_upload_folder=True, we are running this during pre-process
+    # in which case only files in the upload_folder (cache_dir) are accepted
+    if check_in_upload_folder:
+        allowed_paths = []
+    else:
+        allowed_paths = blocks.allowed_paths + [os.getcwd(), tempfile.gettempdir()]
+    allowed, reason = utils.is_allowed_file(
+        abs_path,
+        blocked_paths=blocks.blocked_paths,
+        allowed_paths=allowed_paths,
+        created_paths=created_paths,
+    )
+    if not allowed:
+        msg = f"Cannot move {abs_path} to the gradio cache dir because "
+        if reason == "in_blocklist":
+            msg += f"it is located in one of the blocked_paths ({', '.join(blocks.blocked_paths)})."
+        elif check_in_upload_folder:
+            msg += "it was not uploaded by a user."
+        else:
+            msg += "it was not created by the application or it is not "
+            msg += "located in either the current working directory or your system's temp directory. "
+            msg += "To fix this error, please ensure your function returns files located in either "
+            msg += f"the current working directory ({os.getcwd()}), your system's temp directory ({tempfile.gettempdir()}) "
+            msg += f"or add {str(abs_path.parent)} to the allowed_paths parameter of launch()."
+        raise InvalidPathError(msg)
+    if (
+        utils.is_in_or_equal(abs_path, os.getcwd())
+        and abs_path.name.startswith(".")
+        and not any(
+            is_in_or_equal(path, allowed_path) for allowed_path in blocks.allowed_paths
+        )
+    ):
+        raise InvalidPathError(
+            "Dotfiles located in the temporary directory cannot be moved to the cache for security reasons. "
+            "If you'd like to specifically allow this file to be served, you can add it to the allowed_paths parameter of launch()."
+        )
 
 
 async def async_move_files_to_cache(
@@ -448,14 +589,8 @@ async def async_move_files_to_cache(
             pass
         elif not block.proxy_url:
             # If the file is on a remote server, do not move it to cache.
-            if check_in_upload_folder and not client_utils.is_http_url_like(
-                payload.path
-            ):
-                path = os.path.abspath(payload.path)
-                if not is_in_or_equal(path, get_upload_folder()):
-                    raise ValueError(
-                        f"File {path} is not in the upload folder and cannot be accessed."
-                    )
+            if not client_utils.is_http_url_like(payload.path):
+                _check_allowed(payload.path, check_in_upload_folder)
             if not payload.is_stream:
                 temp_file_path = await block.async_move_resource_to_block_cache(
                     payload.path
@@ -466,10 +601,12 @@ async def async_move_files_to_cache(
                 if keep_in_cache:
                     block.keep_in_cache.add(payload.path)
 
-        url_prefix = "/stream/" if payload.is_stream else "/file="
+        url_prefix = (
+            f"{API_PREFIX}/stream/" if payload.is_stream else f"{API_PREFIX}/file="
+        )
         if block.proxy_url:
             proxy_url = block.proxy_url.rstrip("/")
-            url = f"/proxy={proxy_url}{url_prefix}{payload.path}"
+            url = f"{API_PREFIX}/proxy={proxy_url}{url_prefix}{payload.path}"
         elif client_utils.is_http_url_like(payload.path) or payload.path.startswith(
             f"{url_prefix}"
         ):
@@ -531,7 +668,9 @@ def resize_and_crop(img, size, crop_type="center"):
 ##################
 
 
-def audio_from_file(filename, crop_min=0, crop_max=100):
+def audio_from_file(
+    filename: str, crop_min: float = 0, crop_max: float = 100
+) -> tuple[int, np.ndarray]:
     try:
         audio = AudioSegment.from_file(filename)
     except FileNotFoundError as e:
@@ -544,6 +683,12 @@ def audio_from_file(filename, crop_min=0, crop_max=100):
             else ""
         )
         raise RuntimeError(msg) from e
+    except OSError as e:
+        if wasm_utils.IS_WASM:
+            raise wasm_utils.WasmUnsupportedError(
+                "Audio format conversion is not supported in the Wasm mode."
+            ) from e
+        raise e
     if crop_min != 0 or crop_max != 100:
         audio_start = len(audio) * crop_min / 100
         audio_end = len(audio) * crop_max / 100
@@ -557,6 +702,10 @@ def audio_from_file(filename, crop_min=0, crop_max=100):
 def audio_to_file(sample_rate, data, filename, format="wav"):
     if format == "wav":
         data = convert_to_16_bit_wav(data)
+    elif wasm_utils.IS_WASM:
+        raise wasm_utils.WasmUnsupportedError(
+            "Audio formats other than .wav are not supported in the Wasm mode."
+        )
     audio = AudioSegment(
         data.tobytes(),
         frame_rate=sample_rate,
@@ -782,7 +931,7 @@ def _convert(image, dtype, force_copy=False, uniform=False):
     #   `float32` and `float64` arrays through)
 
     if hasattr(np, "obj2sctype"):
-        is_subdtype = np.issubdtype(dtype_in, np.obj2sctype(dtype))
+        is_subdtype = np.issubdtype(dtype_in, np.obj2sctype(dtype))  # type: ignore
     else:
         is_subdtype = np.issubdtype(dtype_in, dtypeobj_out.type)