+++ js/spa/test/state_change.spec.ts
@@ -56,3 +56,14 @@ test("test datastructure-based state changes", async ({ page }) => {
 	await expect(page.getByLabel("Changes")).toHaveValue("2");
 	await expect(page.getByLabel("Clicks")).toHaveValue("5");
 });
+
+test("test generators properly trigger state changes", async ({ page }) => {
+	await page.getByRole("button", { name: "Iterator State Change" }).click();
+	await expect(page.getByTestId("markdown").first()).toHaveText(
+		"Success Box 0 added"
+	);
+	await page.getByRole("button", { name: "Iterator State Change" }).click();
+	await expect(page.getByTestId("markdown").nth(1)).toHaveText(
+		"Success Box 1 added"
+	);
+});